Data.ConnectionPool

Description

Connection pool used for sharing database connections between threads

Documentation

c :: IO Connection

happyTest :: IO ()

type Pointer = Int

newtype ConnectionContext

A connection context is a numbered connection

Constructors

ConnectionContext

Fields

unConnectionContext :: (Int, Connection)

data ConnectionBucket

A connection bucket can be empty or full If it is empty, it remembers it's connection number

Constructors

Empty Connection Int

Filled Connection

newtype ConnectionPool

The pool is a ring of unused connections and an array of buckets

Constructors

ConnectionPool

Fields

unConnectionPool :: (TQueue Pointer, TArray Int ConnectionBucket)

initConnectionPool :: Int -> IO Connection -> IO ConnectionPool

Startup n connections

emptyConnectionBucket :: ConnectionPool -> Int -> Int -> STM ()

Empty a connection bucket

fillConnectionBucket :: ConnectionPool -> Int -> STM ()

fill a connection bucket

reviveConnection :: ConnectionPool -> ConnectionContext -> IO ConnectionContext

revive a connection (if it is dead)

putConnection :: ConnectionPool -> Int -> Connection -> STM ()

Put a connection back into the pool

getConnection :: ConnectionPool -> IO ConnectionContext

Get a connection from the pool

returnConnection :: ConnectionPool -> ConnectionContext -> IO ()

Return a connection to the pool

unwrapContext :: ConnectionContext -> Connection

Unwrap a connection

unsafeGetConnection :: ConnectionPool -> Int -> STM ConnectionContext

Unsafely get a connection from the pool

initConnectionReclaimer :: ConnectionPool -> Int -> IO ThreadId

This should reclaim a connection after n seconds, but is not yet implemented