Data.InRules

Documentation

hmapKeys :: (Eq k, Hashable k) => (k1 -> k) -> HashMap k1 v -> HashMap k v

hmapWithKey :: (Eq k, Hashable k) => (k -> v1 -> v) -> HashMap k v1 -> HashMap k v

Data types and classes

data InRule

Constructors

InString !String

InByteString !ByteString

InInteger !Integer

InDouble !Double

InNumber !Rational

InBool !Bool

InNull

InArray [InRule]

InObject (HashMap String InRule)

newtype Readable

Constructors

Readable

Fields

unReadable :: String

data InKey

Setters, getters, folds, unfolds and maps.

Constructors

Index Int

None

Assoc String

newtype IdentityMonoid a

Constructors

IM

Fields

unIM :: a

data PathState

Simple automaton for rejecting or accepting paths

Constructors

Accept

Reject

data PathStep a

Constructors

Next (PathAcceptor a)

Final PathState

newtype PathAcceptor a

Constructors

PM

Fields

unPM :: a -> PathStep a

accept :: PathStep a

reject :: PathStep a

acceptor :: PathAcceptor a

Always accept the input

rejector :: PathAcceptor a

Always reject the input

continue :: PathAcceptor a

Always accept the complete input stream (will always be false for finite streams and true for infinite ones)

alter :: PathAcceptor a -> PathAcceptor a -> PathAcceptor a

apoint :: Eq a => a -> PathAcceptor a

runPath :: Eq a => PathAcceptor a -> [a] -> Bool

data KindView

Constructors

TScalar

TArray

TObject

TNone

viewKind :: InRule -> KindView

kmap :: (InKey -> InRule -> InRule) -> InRule -> InRule

Maps through the structure

pmap :: (Monoid (f InKey), Pointed f) => (f InKey -> InRule -> InRule) -> InRule -> InRule

Maps trough the structure with a history of the path kept in a monoid

pfold :: (Monoid (f InKey), Pointed f) => (f InKey -> InRule -> b -> b) -> InRule -> b -> b

Fold trough a structure with a history of the path kept in a monoid

longest_path :: InRule -> Int

Example of the longest path in the inrule structure

ckey :: Num a => [InKey] -> a

kfold :: (InKey -> InRule -> b -> b) -> InRule -> b -> b

Fold through the structure

(.>) :: InRule -> String -> Maybe InRule

Find top level matching keyword

(..>) :: FromInRule a => InRule -> String -> Maybe a

Find top level value and convert to normal value

(.>>) :: InRule -> String -> [InRule]

Search all occuring keywords recursively

readable :: String -> Readable

viaReadable :: Read a => InRule -> a

asReadable :: InRule -> Readable

type InRules = [InRule]

class ToInRule a where

Methods

toInRule :: a -> InRule

class FromInRule a where

Methods

fromInRule :: InRule -> a

toCompatible :: InRule -> InRule

validObject :: InRule -> Bool

emptyObj :: InRule

object :: [(String, InRule)] -> InRule

list :: [InRule] -> InRule

project :: InRule -> InRule -> InRule

mapWithKey :: (k -> a -> b) -> HashMap k a -> HashMap k b

arrayToObj :: InRule -> InRule

shp :: InRule -> InRule -> Bool

shpTestAB :: Bool

shpTestArr :: Bool

singleObj :: ToInRule a => String -> a -> InRule

Create single InRule object.

(==>) :: ToInRule a => String -> a -> InRule

(==>) Eq singleObj .

orM :: Maybe a -> a -> Maybe a

withDefault :: a -> Maybe a -> Maybe a

fromList :: ToInRule a => [(String, a)] -> InRule

Create InRule object from list.

toList :: FromInRule a => InRule -> [(String, a)]

Create InRule object from list.

toListString :: InRule -> [(String, String)]

unionObj :: InRule -> InRule -> InRule

unionsObj :: [InRule] -> InRule

Merge InRule objects from list.

unionRecObj :: InRule -> InRule -> InRule

toString :: InRule -> String

Renders InRule to String.

Pretty print

pprint :: InRule -> IO ()

Pretty-prints InRule.

pprint' :: String -> Integer -> InRule -> String

pprints :: [InRule] -> IO ()

Pretty-prints InRules.

escInStr :: String -> String

escInChar :: Char -> String