Data.MemTimeState

Documentation

data MemState

The Map is a mapping from a bytestring key to the internal used key. | MemMap contains the binary data, indexed by the internal key | TimeMap contains the internal keys index by time

data Query where

Constructors

Insert :: Key -> Value -> Query

Delete :: Key -> Query

Query :: Key -> Query

DumpState :: Query

data Result where

Constructors

Value :: ByteString -> Result

NotFound :: Result

Empty :: Result

Except :: String -> Result

KeyVal :: ByteString -> ByteString -> Result

type QueryChan = TQueue (Query, Unique Result)

runQuery :: QueryChan -> Query -> IO Result

newMemState :: Time -> Time -> FilePath -> IO MemState

queryManager :: FilePath -> MemState -> QueryChan -> IO ()