Data.PriorityQueue

Description

This is a reasonable implementation of a priority queue. It has O(1) lookup and amortized delete min of O(log n) I am to lazy to calculate insert, but in practice it works quite well.

Documentation

data Prio q a

Priority heap, the priority is a parameter to the type Prio q a can be ordered, is a functor, a foldable, a monoid and a traversable. There is also a monadic version. We should interpret this as a computation on the objects in the priority heap, rendering a tree of priority heaps and then merged into a consistent priority heap again.

The monad has a problem, because it has a constraint. In the current version of ghc, it is not possible to make it a monad yet.

A comonad like instance could be useful, but is not possible, because a comonad should always be safe to deconstruct. This thing is clearly not safe to deconstruct.

It could be empty

view :: Ord q => Prio q a -> ViewMin q a

View pattern for list like perspective

data ViewMin q a

List like view for the minimal element

Constructors

a :> (Prio q a)

Nil

fromList :: Ord q => [(q, a)] -> Prio q a

Build a prio heap from a list

extractMin :: Ord q => Prio q a -> Maybe (a, Prio q a)

Get the minimal value O(1)

headMin :: Ord q => Prio q a -> Maybe a

Get the first element

tailMin :: Ord q => Prio q a -> Maybe (Prio q a)

Get the tail

insert :: Ord q => q -> a -> Prio q a -> Prio q a

Insert a new element with priority

singleton :: q -> a -> Prio q a

Create an singleton prio queue

extractTill :: Ord q => q -> Prio q a -> ([a], Prio q a)

Get all till priority as value

extractTillWithKey :: Ord q => q -> Prio q a -> ([(q, a)], Prio q a)

Get all till priority as tuple