Data.Role

Documentation

type RolePair k = (TimedMap k [Role], RoleSetMap)

data RoleSetFile

Constructors

Roles [String] [RoleSetFile]

Resource String [String]

data RestRight

Constructors

Put

Get

Post

Delete

newtype RoleSetMap

Constructors

RSM

Fields

unRSM :: TVar RoleSet

type Resource = String

type Id = Integer

data Role

Constructors

Developer (Maybe Id)

User (Maybe Id)

Server (Maybe Id)

All

data OpaqueRole

Constructors

ODeveloper

OUser

OServer

OAll

viewOpaque :: Role -> OpaqueRole

newtype RoleSet

Constructors

RS

Fields

unRS :: Map Resource (Map Role [RestRight])

initRP :: MonadIO m => FilePath -> m (RolePair ByteString)

fileStore :: FilePath -> TimedMapStore ByteString [Role]

File store handler

fileRestore :: FilePath -> TimedMapRestore ByteString [Role]

File restore handler

voidStore :: Ord k => TimedMapStore k a

Empty Store handler

voidRestore :: Ord k => TimedMapRestore k a

Empty Restore handler

getRoles :: MonadIO m => RolePair ByteString -> ByteString -> m [Role]

Retrieve all the roles from a RoleState from the given token.

updateTimeRole :: MonadIO m => RolePair ByteString -> ByteString -> m ()

may :: MonadIO m => RolePair ByteString -> Resource -> Role -> RestRight -> m Bool

Lookup from the RoleState if a user may access a resource as role with restright

addRole :: MonadIO m => RolePair ByteString -> ByteString -> Role -> m ()

Adds a role to the RoleState under the given token.

dropRoles :: MonadIO m => RolePair ByteString -> ByteString -> m ()

Drop all the roles from token in the RoleState

runCleanup :: RolePair ByteString -> Int64 -> IO ()

Cleanup all expired tokens

initCleanup :: RolePair ByteString -> Int64 -> IO ThreadId

Start separated thread to cleanup all tokens wich are expired

initStore :: RolePair ByteString -> TimedMapStore ByteString [Role] -> IO ThreadId

Start the storing thread, which periodically stores the whole RoleState with the given handler

runRestore :: RolePair ByteString -> TimedMapRestore ByteString [Role] -> IO ()

Restore the roleState from the given handler

dumpAll :: MonadIO m => RolePair ByteString -> m String

Debug function to dump the internal state

may' :: RoleSet -> Resource -> Role -> RestRight -> Bool

Pure function to lookup from a RoleSet if a Role can acces a resource under the given rights

roleSet :: Parser RoleSetFile

roleResources :: Parser [(String, [String])]

roleHeader :: Parser [String]

resourceRule :: Parser (String, [String])

hsep :: Parser Char

defs :: Parser String

vsepB :: Parser Char

vsepH0 :: Parser String

vsepH :: Parser String

roles :: Parser String

identifier :: Parser Char

zipRoleSet :: RoleSetFile -> RoleSet

readRole :: String -> Role