Data.SqlTransaction

Documentation

quickInsert :: IConnection c => String -> [(String, SqlValue)] -> SqlTransaction c SqlValue

quickInsert: insert data map into a single specified table. data map has the form [(field, value)]. values are SqlValues. lastval() is returned.

data Connection

class IConnection conn where

Methods

disconnect :: conn -> IO ()

data SqlValue

Constructors

SqlString String

SqlByteString ByteString

SqlWord32 Word32

SqlWord64 Word64

SqlInt32 Int32

SqlInt64 Int64

SqlInteger Integer

SqlChar Char

SqlBool Bool

SqlDouble Double

SqlRational Rational

SqlLocalDate Day

SqlLocalTimeOfDay TimeOfDay

SqlZonedLocalTimeOfDay TimeOfDay TimeZone

SqlLocalTime LocalTime

SqlZonedTime ZonedTime

SqlUTCTime UTCTime

SqlDiffTime NominalDiffTime

SqlPOSIXTime POSIXTime

SqlEpochTime Integer

SqlTimeDiff Integer

SqlNull

data Statement

disconnect :: IConnection conn => conn -> IO ()

data Lock

Several locks in the postgresql database

Constructors

AccessShare

RowShare

RowExclusive

ShareUpdateExclusive

Share

ShareRowExclusive

AccessExclusive

type SqlTransaction c a = SqlTransactionUser Lock c a

Sql transaction is a monad, which encapsulate computations in a transaction By throwing an error the transaction will be rolled back The SqlTransactionUser gives the possibility to add a user state In SqlTransaction the user state is a mechanism to lock a certain operation Other capabilities are:

paralel queries

atomical actions

exception catching (Error)

error handling

server level locking

The monad is hand rolled and CPS transformed for efficiency

data SqlTransactionUser l c a

This is the core monad. It is a hand rolled CPS transformed monadstack. It handles state and exception handling

It is derived from the following stack:

   newtype SqlTransaction l c a = SqlTransaction {
     unsafeRunSqlTransaction :: StateT (c,l) (ErrorT String IO) a 
   } deriving (Functor, Alternative, Applicative, Monad, MonadPlus, MonadFix, MonadState c, MonadError String, MonadIO) 

atomical :: SqlTransaction Connection a -> SqlTransaction Connection a

Explicitly encapsulates a computation in a transaction block. This commits the previous computation

catchSqlError :: IConnection c => SqlTransactionUser l c a -> (SqlError -> SqlTransactionUser l c a) -> SqlTransactionUser l c a

Catch an error thrown in the SqlTransaction monad, if there is an error the continuation provided by the user will be run

commit :: SqlTransaction Connection ()

Commits the current data to the database and starts a new transaction

dbWithLockBlock :: Show a => Namespace -> a -> SqlTransactionUser Lock c b -> SqlTransactionUser Lock c b

Do a SqlTransaction action, lock on server level only blocks until lock is acquired

dbWithLockNonBlock :: Show a => Namespace -> a -> SqlTransactionUser Lock c () -> SqlTransactionUser Lock c ()

Do a SqlTransaction action and put a lock on the provided label. If the lock can't be acquired, don't block.

Example

 withLockNonBlock namespace key $ do ... 

doneFuture :: Future a -> SqlTransaction c Bool

Check if the future is done calculating

emptyFuture :: SqlTransaction c (Future a)

Create an empty future

execute :: IConnection c => Statement -> [SqlValue] -> SqlTransaction c Integer

executeMany :: IConnection c => Statement -> [[SqlValue]] -> SqlTransaction c ()

executeRaw :: IConnection c => Statement -> SqlTransaction c ()

fetchAllRows' :: IConnection c => Statement -> SqlTransaction c [[SqlValue]]

fetchAllRows :: IConnection c => Statement -> SqlTransaction c [[SqlValue]]

fetchAllRowsAL' :: IConnection c => Statement -> SqlTransaction c [[(String, SqlValue)]]

fetchAllRowsAL :: IConnection c => Statement -> SqlTransaction c [[(String, SqlValue)]]

fetchAllRowsMap :: IConnection c => Statement -> SqlTransaction c [Map String SqlValue]

fetchRow :: IConnection c => Statement -> SqlTransaction c (Maybe [SqlValue])

fetchRowAl :: IConnection c => Statement -> SqlTransaction c (Maybe [(String, SqlValue)])

fetchRowMap :: IConnection c => Statement -> SqlTransaction c (Maybe (Map String SqlValue))

fillFuture :: Future a -> Either SqlError a -> SqlTransaction c ()

Fill a future with a value This is an internal function

finish :: Statement -> SqlTransaction c ()

forkSqlTransaction :: SqlTransaction Connection () -> SqlTransaction Connection ThreadId

Fork a SqlTransaction, so it can calculate the computation concurrently This shares the database connection, so it the parent should be done with all the operations on the database. Errors don't roll back the parent

get :: MonadState s m => m s

getUser :: SqlTransactionUser l c l

Retrieve the user state

lock :: String -> Lock -> SqlTransaction Connection a -> SqlTransaction Connection a

Lock a table with Lock and do a computation when locked

Example

 f = lock account RowExclusive $ do ... 

newFuture

Arguments

:: IConnection c

=> SqlTransaction c a
calculation needed in the future

-> SqlTransaction c (Future a)

Creates a new future in the SqlTransaction monad with a calculation The database connection will be cloned, so it is safe for the parent to operate on the database

par2 :: IConnection c => SqlTransaction c a -> SqlTransaction c b -> SqlTransaction c (a, b)

Run two computations paralel

par3 :: IConnection c => SqlTransaction c a -> SqlTransaction c b -> SqlTransaction c c -> SqlTransaction c (a, b, c)

Run three computations paralel

par4 :: IConnection c => SqlTransaction c p -> SqlTransaction c q -> SqlTransaction c r -> SqlTransaction c s -> SqlTransaction c (p, q, r, s)

Run four computations in paralel

parN :: IConnection c => [SqlTransaction c p] -> SqlTransaction c [p]

run n computations in paralel

parSafe :: [SqlTransactionUser Lock Connection b] -> SqlTransactionUser Lock Connection [b]

Run queries in paralel Example:

parSafe [comp1, comp2, comp3] This will return all the results or roll back the parent computation on an error

prepare :: IConnection c => String -> SqlTransaction c Statement

putUser :: l -> SqlTransactionUser l c ()

Put the user state

quickQuery' :: IConnection c => String -> [SqlValue] -> SqlTransaction c [[SqlValue]]

quickQuery :: IConnection c => String -> [SqlValue] -> SqlTransaction c [[SqlValue]]

readFuture :: IConnection c => Future a -> SqlTransaction c a

Read the future, this will force the calculation. Any exception will be thrown in the parent

rollback :: IConnection m => String -> SqlTransaction m t

Rollback the computation, throws an user error

run :: IConnection c => String -> [SqlValue] -> SqlTransaction c Integer

runSqlTransaction :: (Applicative m, MonadIO m, IConnection conn) => SqlTransactionUser l conn b -> (String -> m b) -> conn -> l -> m b

Encapsulates the computation in a transaction and handles any errors in the user supplied function

runTestDb :: SqlTransactionUser l Connection b -> IO b

sExecute :: IConnection c => Statement -> [Maybe String] -> SqlTransaction c Integer

sExecuteMany :: IConnection c => Statement -> [[Maybe String]] -> SqlTransaction c ()

sFetchAllRows' :: IConnection c => Statement -> SqlTransaction c [[Maybe String]]

sFetchAllRows :: IConnection c => Statement -> SqlTransaction c [[Maybe String]]

sFetchRow :: Statement -> SqlTransaction c (Maybe [Maybe String])

sRun :: IConnection c => String -> [Maybe String] -> SqlTransaction c Integer

data SqlError

Exception data type.

Constructors

UError String
User error

DBError String
Database error

sqlExecute :: IConnection c => String -> [SqlValue] -> SqlTransaction c ()

Execute a statement

sqlGetAll :: IConnection c => String -> [SqlValue] -> SqlTransaction c [[SqlValue]]

Get all rows from the query

sqlGetAllAssoc :: IConnection c => String -> [SqlValue] -> SqlTransaction c [HashMap String SqlValue]

Get all rows as a HashMap

sqlGetOne :: IConnection c => String -> [SqlValue] -> SqlTransaction c SqlValue

Get the first value from the query

sqlGetRow :: IConnection c => String -> [SqlValue] -> SqlTransaction c [SqlValue]

get one row from the query

waitUnless :: SqlTransaction Connection Bool -> SqlTransaction Connection ()

The reverse of waitWhen

waitWhen :: SqlTransaction Connection Bool -> SqlTransaction Connection ()

Wait on an event in the database. If the computation returns true, go further

withEncoding :: IConnection c => String -> SqlTransaction c a -> SqlTransaction c a

Change the encoding of the database