Data.TimedMap

Documentation

newtype TimedMap k a

Constructors

TimedMap

Fields

unTimeMap :: TVar (TMap k a)

type TimedMapRestore k a = IO [(Int64, k, a)]

type TimedMapStore k a = [(Int64, k, a)] -> IO ()

newtype TMap k a

Constructors

TMap

Fields

unTMap :: (Map k Int64, Map k a)

newTimedMap :: Ord k => IO (TimedMap k a)

creates a new timed map

getTimeStamp :: IO Int64

Helper function, gets a time stamp as a 64 bits integer

withNow0 :: (a -> Int64 -> IO b) -> a -> IO b

Some helper functions to work with the now

withNow1 :: (a -> Int64 -> b -> IO c) -> a -> b -> IO c

withNow2 :: (a -> Int64 -> b -> c -> IO d) -> a -> b -> c -> IO d

withNow3 :: (a -> Int64 -> b -> c -> d -> IO e) -> a -> b -> c -> d -> IO e

lookup :: Ord k => TimedMap k a -> k -> STM (Maybe a)

Lookup a value from a TimedMap

lookupTime :: Ord k => TimedMap k a -> k -> STM (Maybe Int64)

Lookup a time from a TimedMap

lookupBoth' :: Ord k => TimedMap k a -> k -> STM (Maybe (Int64, a))

Lookup a time and other value from a TimedMap

tupleMaybe :: (Maybe Int64, Maybe b) -> Maybe (Int64, b)

lookupBoth :: Ord k => TimedMap k a -> k -> STM (Maybe Int64, Maybe a)

Lookup a time and other value from a TimedMap

insert :: Ord k => TimedMap k a -> Int64 -> k -> a -> STM ()

Insert a value in a TimedMap

cleanup :: Ord k => TimedMap k a -> Int64 -> STM ()

Cleanup old variables from a TimedMap, the expiration time is given

updateTime :: Ord k => TimedMap k a -> Int64 -> k -> STM ()

Update the time from a key

delete :: Ord k => TimedMap k a -> k -> STM ()

Delete a key from a TimedMap

tlookup :: Ord k => TMap k a -> k -> Maybe a

tlookupTime :: Ord k => TMap k a -> k -> Maybe Int64

tlookupBoth :: Ord k => TMap k a -> k -> (Maybe Int64, Maybe a)

tinsert :: Ord k => TMap k a -> Int64 -> k -> a -> TMap k a

tupdateTime :: Ord k => TMap k a -> Int64 -> k -> TMap k a

tcleanup :: Ord k => TMap k a -> Int64 -> TMap k a

tremove :: Ord a => Int64 -> Map a Int64 -> (Map a Int64, [a])

tdelete :: Ord k => TMap k a -> k -> TMap k a

storeTimedMap :: Ord k => TimedMapStore k a -> TimedMap k a -> IO ()

restoreTimedMap :: Ord k => TimedMap k a -> TimedMapRestore k a -> IO ()

This is not an atomically safe action! Restores the map with the given handler. An insert to the map will be overwritten

storeTMap :: Ord k => TimedMapStore k a -> TMap k a -> IO ()

restoreTMap :: Ord k => TimedMapRestore k a -> IO (TMap k a)